+++ examples/train_ticket2/workflow/train_ticket2/food_data.go
@@ -8,7 +8,7 @@ type Food struct {
 type FoodOrder struct {
 	ID          string
 	OrderID     string
-	FoodType    int64
+	FoodType    int
 	StationName string
 	StoreName   string
 	FoodName    string
